core-refactor: fix import alias
