+++ packages/core/src/dom/children/switch.ts
@@ -1,9 +1,9 @@
 import { Component } from '@component/componentBlock.ts'
 import { componentContext } from '@context/executionContext.ts'
-import { Block } from '@dom/type.ts'
 import { GetState, isGetState } from '@hook/useState.ts'
 import { isRvjsFunction } from '@type/guard.ts'
 import { RvjsFunction } from '@type/rvjs.ts'
+import { Block } from '@type/type.ts'
 import { Context } from '@util/context.ts'
 import { RVJS_SWITCH_RENDER_SYMBOL } from '@util/symbol.ts'
