+++ packages/core/src/dom/element/elementBlock.ts
@@ -3,8 +3,8 @@ import { isSwitchRender } from '@children/switch.ts'
 import { isToggleRender, ToggleRender } from '@children/toggle.ts'
 import { isComponent } from '@component/componentBlock.ts'
 import { subscribeStateContext } from '@context/executionContext.ts'
-import { Block, Children, Render } from '@dom/type.ts'
 import { isArray, isTextNode } from '@type/guard.ts'
+import { Block, Children, Render } from '@type/type.ts'
 import { insertChildrenAtIndex } from '@util/dom.ts'
 import { Observer } from '@util/observer.ts'
 import { RVJS_ELEMENT_SYMBOL } from '@util/symbol.ts'
