+++ packages/core/src/dom/index.ts
@@ -1,15 +1,15 @@
-export { For } from '@children/for'
-export { Switch } from '@children/switch'
-export { Toggle } from '@children/toggle'
-export { Suspense } from '@children/suspense'
-export { component } from '@component/component'
-export { root } from '@component/root'
-export * from '@element/elementMap'
+export { For } from '@children/for.ts'
+export { Switch } from '@children/switch.ts'
+export { Toggle } from '@children/toggle.ts'
+export { Suspense } from '@children/suspense.ts'
+export { component } from '@component/component.ts'
+export { root } from '@component/root.ts'
+export * from '@element/elementMap.ts'
 export { overrideElement } from '@element/util.ts'
 export { Component } from '@component/componentBlock.ts'
 export { Element } from '@element/elementBlock.ts'
 export { isTextNode } from '@type/guard.ts'
 export { isRvjsFunction, isRvjsObject } from '@type/guard.ts'
-export type { ComponentFn } from '@component/component'
-export type * from '@dom/type.ts'
+export type { ComponentFn } from '@component/component.ts'
+export type * from '@type/type.ts'
 export type * from '@element/type.ts'
