+++ packages/core/src/reactive/context/executionContext.ts
@@ -1,7 +1,7 @@
 import { Component } from '@component/componentBlock.ts'
-import { Block } from '@dom/type.ts'
 import { Dynamic } from '@hook/dynamic.ts'
 import { GetState } from '@hook/useState.ts'
+import { Block } from '@type/type.ts'
 import { Context } from '@util/context.ts'
 
 export interface StateContext {
