+++ packages/core/src/reactive/hook/useElement.ts
@@ -1,6 +1,6 @@
 import { isComponent } from '@component/componentBlock.ts'
-import { Block } from '@dom/type.ts'
 import { isElement } from '@element/elementBlock.ts'
+import { Block } from '@type/type.ts'
 
 export const useElement = (block: Block) => {
   if (isElement(block)) {
