+++ packages/core/src/reactive/index.ts
@@ -1,13 +1,13 @@
-export { dynamic } from '@hook/dynamic'
-export { useEffect } from '@hook/useEffect'
-export { useRef } from '@hook/useRef'
-export { useState, isGetState } from '@hook/useState'
-export { useGlobalState } from '@hook/useGlobalState'
-export { onMount } from '@lifecycle/onMount'
-export { onDestroy } from '@lifecycle/onDestroy'
+export { dynamic } from '@hook/dynamic.ts'
+export { useEffect } from '@hook/useEffect.ts'
+export { useRef } from '@hook/useRef.ts'
+export { useState, isGetState } from '@hook/useState.ts'
+export { useGlobalState } from '@hook/useGlobalState.ts'
+export { onMount } from '@lifecycle/onMount.ts'
+export { onDestroy } from '@lifecycle/onDestroy.ts'
 export { useElement } from '@hook/useElement.ts'
 export { prop } from '@hook/prop.ts'
-export type { RefObject } from '@hook/useRef'
-export type { GetState, SetState } from '@hook/useState'
-export type { Dynamic } from '@hook/dynamic'
+export type { RefObject } from '@hook/useRef.ts'
+export type { GetState, SetState } from '@hook/useState.ts'
+export type { Dynamic } from '@hook/dynamic.ts'
 export type { Prop } from '@hook/prop.ts'
