+++ packages/core/src/router/hook/useOutlet.ts
@@ -1,7 +1,7 @@
 import { Switch } from '@children/switch.ts'
 import { componentContext } from '@context/executionContext.ts'
-import { Block } from '@dom/type.ts'
 import { useState } from '@hook/useState.ts'
+import { Block } from '@type/type.ts'
 
 const useOutlet = () => {
   const [outlet, setOutlet] = useState<Block | null>(null)
