+++ packages/core/src/router/index.ts
@@ -1,8 +1,8 @@
-export { default as Router } from '@router/router'
-export { default as Link } from '@router/link'
-export { default as useNavigate } from '@router/hook/useNavigate'
-export { default as useOutlet } from '@router/hook/useOutlet'
-export { default as usePathEffect } from '@router/hook/usePathEffect'
-export { default as usePathname } from '@router/hook/usePathname'
-export { default as usePathParams } from '@router/hook/usePathParams'
-export { default as useQueryParams } from '@router/hook/useQueryParams'
+export { default as Router } from '@router/router.ts'
+export { default as Link } from '@router/link.ts'
+export { default as useNavigate } from '@router/hook/useNavigate.ts'
+export { default as useOutlet } from '@router/hook/useOutlet.ts'
+export { default as usePathEffect } from '@router/hook/usePathEffect.ts'
+export { default as usePathname } from '@router/hook/usePathname.ts'
+export { default as usePathParams } from '@router/hook/usePathParams.ts'
+export { default as useQueryParams } from '@router/hook/useQueryParams.ts'
