+++ packages/core/src/util/index.ts
@@ -1,1 +1,1 @@
-export * from './symbol'
+export * from '@util/symbol.ts'
